requirements.txt
# Django & REST
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5
drf-spectacular>=0.27  # OpenAPI/Swagger

# Auth & Security
djangorestframework-simplejwt>=5.3
django-otp>=1.2
qrcode>=7.4
cryptography>=41.0
django-ratelimit>=4.1

# Celery & Redis
celery[redis]>=5.3
redis>=5.0
django-celery-beat>=2.5

# Database
psycopg2-binary>=2.9

# File handling & Reports
openpyxl>=3.1
xlsxwriter>=3.1
reportlab>=4.0
Pillow>=10.0
django-cleanup>=8.0

# Notifications & Channels (WebSockets)
channels>=4.0
channels-redis>=4.2
daphne>=4.0

# Utils
python-dateutil>=2.8
python-dotenv>=1.0
gunicorn>=21.0
whitenoise>=6.6

# Testing
pytest>=7.4
pytest-django>=4.7
pytest-cov>=4.1
factory-boy>=3.3
httpx>=0.25

# Code quality
black>=23.12
ruff>=0.1
